Redesigned workspace algorithmization system. Tool info code now has the float type.

+++ Sources/IndustrialKit/Connector.swift
@@ -329,7 +329,7 @@ open class ToolConnector: WorkspaceObjectConnector
     }
     
     ///Inforamation code updated by connector.
-    public var info_code: Int?
+    public var info_code: Float?
     
     ///A tool model controller.
     public var model_controller: ToolModelController?
